requirements.txt
APScheduler==3.9.1
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==3.2.13
django-apscheduler==0.6.2
djangorestframework==3.13.1
install==1.3.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
